requirements.txt
absl-py==2.3.1
albumentations==1.1.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7.1
fastprogress==1.0.3
filelock==3.19.1
Flask==3.1.2
fsspec==2025.7.0
gensim==4.3.3
grpcio==1.74.0
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
Markdown==3.8.2
MarkupSafe==3.0.2
MolScribe==1.1.1
networkx==3.2.1
numpy==1.23.5
opencv-python==4.11.0.86
opencv-python-headless==4.6.0.66
OpenNMT-py==2.2.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.0
PubChemPy==1.0.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qudida==0.0.4
rdkit-pypi==2022.9.5
requests==2.32.5
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.2.1
six==1.17.0
smart_open==7.3.0.post1
SmilesPE==0.0.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2024.8.30
timm==0.4.12
torch==1.13.1
torchaudio==0.13.1
torchtext==0.5.0
torchvision==0.14.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.3
zipp==3.23.0
